requirements.txt
openai>=2.24.0
python-dotenv>=1.2.2
